Move ContentManager interface to protocol package
