--- org-code-javabuilder/protocol/src/main/java/org/code/javabuilder/ContentManager.java
@@ -1,12 +1,0 @@
-package org.code.javabuilder;
-
-import org.code.protocol.JavabuilderException;
-
-public interface ContentManager {
-    byte[] getAssetData(String filename) throws JavabuilderException;
-
-    String generateAssetUploadUrl(String filename) throws JavabuilderException;
-
-    String writeToOutputFile(String filename, byte[] inputBytes, String contentType)
-            throws JavabuilderException;
-}
